+++ frontend/app/routes/services/settings/services-settings.tsx
@@ -33,19 +33,19 @@ import {
 } from "~/lib/queries";
 import { cn } from "~/lib/utils";
 import { queryClient } from "~/root";
-import { ServiceBuilderForm } from "~/routes/services/settings/service-builder-form";
-import { ServiceCommandForm } from "~/routes/services/settings/service-command-form";
-import { ServiceConfigsForm } from "~/routes/services/settings/service-configs-form";
-import { ServiceDangerZoneForm } from "~/routes/services/settings/service-danger-zone-form";
-import { ServiceDeployURLForm } from "~/routes/services/settings/service-deploy-url-form";
-import { ServiceGitSourceForm } from "~/routes/services/settings/service-git-source-form";
-import { ServiceHealthcheckForm } from "~/routes/services/settings/service-healthcheck-form";
-import { ServicePortsForm } from "~/routes/services/settings/service-ports-form";
-import { ServiceResourceLimits } from "~/routes/services/settings/service-resource-limits-form";
-import { ServiceSlugForm } from "~/routes/services/settings/service-slug-form";
-import { ServiceSourceForm } from "~/routes/services/settings/service-source-form";
-import { ServiceURLsForm } from "~/routes/services/settings/service-urls-form";
-import { ServiceVolumesForm } from "~/routes/services/settings/service-volumes-form";
+import { ServiceBuilderForm } from "~/routes/services/components/service-builder-form";
+import { ServiceCommandForm } from "~/routes/services/components/service-command-form";
+import { ServiceConfigsForm } from "~/routes/services/components/service-configs-form";
+import { ServiceDangerZoneForm } from "~/routes/services/components/service-danger-zone-form";
+import { ServiceDeployURLForm } from "~/routes/services/components/service-deploy-url-form";
+import { ServiceGitSourceForm } from "~/routes/services/components/service-git-source-form";
+import { ServiceHealthcheckForm } from "~/routes/services/components/service-healthcheck-form";
+import { ServicePortsForm } from "~/routes/services/components/service-ports-form";
+import { ServiceResourceLimits } from "~/routes/services/components/service-resource-limits-form";
+import { ServiceSlugForm } from "~/routes/services/components/service-slug-form";
+import { ServiceSourceForm } from "~/routes/services/components/service-source-form";
+import { ServiceURLsForm } from "~/routes/services/components/service-urls-form";
+import { ServiceVolumesForm } from "~/routes/services/components/service-volumes-form";
 import { getCsrfTokenHeader, wait } from "~/utils";
 import { type Route } from "./+types/services-settings";
